Add filter for midonet networking types
In the midonet ML2 mechanism driver, we want to only notify the
midonet backend of the changes happening on networks, ports, and
subnets that belong to networks with the 'midonet' network type.
This patch adds a decorator that filters function calls based
on the network type.

The filtering for each resource type (port, subnet, and network)
are each different, so different ways of getting the network
are needed for each.

Change-Id: I6c0ebc75b77ece0e96ad6ee152fb62be2ce52afe
Signed-off-by: Joe Mills <joe@midokura.com>